requirements-dev.txt
#basic dependencies for development
pydantic
fastapi
uvicorn
python-dotenv
pymongo
logging
typing
typing_extensions
fastapi-mail